Moved all test-functions into test.cpp
