--- include/test_matrix_operations.hpp
@@ -1,8 +1,0 @@
-#ifndef __test_matrix_operations__
-#define __test_matrix_operations__
-
-#include <armadillo>
-
-void test_max_offdiag_symmetric();
-
-#endif
